+++ src/test/universal/graphqlFetchOptions.test.mjs
@@ -44,12 +44,7 @@ export default (tests) => {
     ]);
     deepStrictEqual(formDataEntries[1], ['map', '{"1":["variables.a"]}']);
     strictEqual(formDataEntries[2][0], '1');
-    strictEqual(typeof formDataEntries[2][1], 'object');
-
-    // Sadly `instanceof Blob` won’t work, see:
-    // https://github.com/octet-stream/form-data/issues/14
-    strictEqual(formDataEntries[2][1].constructor.name, 'File');
-
+    strictEqual(formDataEntries[2][1] instanceof Blob, true);
     strictEqual(formDataEntries[2][1].name, 'blob');
     strictEqual(formDataEntries[2][1].type, filetype);
   });